+++ RELEASENOTES.md
@@ -63,6 +63,9 @@
         a notification update
         ([#1833](https://github.com/androidx/media/issues/1833)).
     *   Add `ProgressListener` to custom command methods.
+    *   Change the default value for
+        `MediaLibrarySession.Builder.setLibraryErrorReplicationMode` to non
+        fatal.
 *   UI:
     *   Add `ProgressStateWithTickInterval` class and the corresponding
         `rememberProgressStateWithTickInterval` Composable to